Move SortableGrid
