+++ core/src/com/unciv/ui/screens/overviewscreen/CityOverviewTab.kt
@@ -1,7 +1,7 @@
 package com.unciv.ui.screens.overviewscreen
 
 import com.unciv.logic.civilization.Civilization
-import com.unciv.ui.components.SortableGrid
+import com.unciv.ui.components.widgets.SortableGrid
 import com.unciv.ui.components.extensions.equalizeColumns
